backend/pyproject.toml
[project]
name = "dockyy-backend"
version = "0.1.0"
description = "Docker dashboard backend API"
requires-python = ">=3.11"
dependencies = [
  "flask>=3.0.0",
  "flask-cors>=4.0.0",
  "docker>=7.0.0",
  "python-dotenv>=1.0.0",
  "sqlalchemy-libsql>=0.2.0",
  "alembic>=1.17.2",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "black>=23.0.0", "flake8>=6.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
